eventroop_backend/requirements.txt
asgiref==3.10.0
boto3==1.40.58
botocore==1.40.58
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
jmespath==1.0.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
